refactor: move loader to state and rename game
Prior to this change, I had mixed up the idea of states and systems and
had mistakenly classified Loader as a system. It is a state, and using
it just for the loading step is probably not a good design.

This change moves the loader to a new directory called state, which is
where game states will go. It renames it to game.rs, and it will now
represent the main game state. It still does loading in its on_start,
this might end up getting moved elsewhere, like setup for the various
systems, which might help with the issue currently with systems running
before all the assets are loaded.